+++ compare-frontend/src/app/state/reducer.ts
@@ -4,8 +4,9 @@ import { WritableDraft } from 'immer/dist/internal';
 import { RiesgosState, initialRiesgosState, RiesgosProduct, RiesgosStep, ScenarioName, StepStateAvailable, StepStateCompleted, StepStateTypes, StepStateUnavailable, StepStateRunning, StepStateError, Partition, RiesgosScenarioState } from './state';
 import { ruleSetPicked, scenarioLoadStart, scenarioLoadSuccess, scenarioLoadFailure, stepSetFocus, stepConfig, stepExecStart, stepExecSuccess, stepExecFailure, scenarioPicked, autoPilotStart, autoPilotStop, autoPilotDequeue, autoPilotEnqueue, mapMove, mapClick, togglePartition } from './actions';
 import { API_ScenarioInfo } from '../services/backend.service';
-import { allParasSet } from './helpers';
+import { allParasSet, getMapPositionForStep } from './helpers';
 import { TypedAction } from '@ngrx/store/src/models';
+import { act } from '@ngrx/effects';
 
 
 
@@ -44,22 +45,24 @@ export const reducer = createReducer(
     const scenarioData = state.scenarioData[action.scenario]!;
     const partitionData = scenarioData[action.partition]!;
 
-    if (action.focus === false) {
-      partitionData.focus.focusedSteps = partitionData.focus.focusedSteps.filter(s => s !== action.stepId);
-    } else {
-      if (state.rules.oneFocusOnly) partitionData.focus.focusedSteps = [action.stepId];
-      else partitionData.focus.focusedSteps.push(action.stepId);
+    function handle(pd: RiesgosScenarioState, action: ReturnType<typeof stepSetFocus>) {
+      if (action.focus === false) {
+        pd.focus.focusedSteps = pd.focus.focusedSteps.filter(s => s !== action.stepId);
+      } else {
+        if (state.rules.oneFocusOnly) pd.focus.focusedSteps = [action.stepId];
+        else if (!pd.focus.focusedSteps.includes(action.stepId)) pd.focus.focusedSteps.push(action.stepId);
+        const {center, zoom} = getMapPositionForStep(action.scenario, action.partition, action.stepId);
+        pd.map.center = center;
+        pd.map.zoom = zoom;
+      }
     }
 
     if (state.rules.mirrorFocus) {
       for (const [otherPartition, otherPartitionData] of Object.entries(scenarioData)) {
-        if (action.focus === false) {
-          otherPartitionData.focus.focusedSteps = otherPartitionData.focus.focusedSteps.filter(s => s !== action.stepId);
-        } else {
-          if (state.rules.oneFocusOnly) otherPartitionData.focus.focusedSteps = [action.stepId];
-          else otherPartitionData.focus.focusedSteps.push(action.stepId);
-        }
+        handle(otherPartitionData, action);
       }
+    } else {
+      handle(partitionData, action);
     }
 
     return state;
@@ -70,24 +73,22 @@ export const reducer = createReducer(
     const scenarioData = state.scenarioData[action.scenario]!;
     const partitionData = scenarioData[action.partition]!;
 
-    for (const [productId, productValue] of Object.entries(action.values)) {
-      for (const product of partitionData.products) {
-        if (product.id === productId) {
-          product.value = productValue;
+    function handle(pd: RiesgosScenarioState, action: ReturnType<typeof stepConfig>) {
+      for (const [productId, productValue] of Object.entries(action.values)) {
+        for (const product of pd.products) {
+          if (product.id === productId) {
+            product.value = productValue;
+          }
         }
       }
     }
 
     if (state.rules.mirrorData) {
       for (const [otherPartition, otherPartitionData] of Object.entries(scenarioData)) {
-        for (const [productId, productValue] of Object.entries(action.values)) {
-          for (const product of otherPartitionData.products) {
-            if (product.id === productId) {
-              product.value = productValue;
-            }
-          }
-        }  
+        handle(otherPartitionData, action);
       }
+    } else {
+      handle(partitionData, action);
     }
 
     return state;
@@ -227,8 +228,8 @@ export const reducer = createReducer(
 
 
   immerOn(togglePartition, (state, action) => {
-    const scenarioState = state.scenarioData[action.scenario]![action.partition]!;
-    scenarioState.active = !scenarioState.active;
+    const partitionData = state.scenarioData[action.scenario]![action.partition]!;
+    partitionData.active = !partitionData.active;
     return state;
   })
 );